feat:reflectionの認可追加

+++ app/controllers/reflections_controller.rb
@@ -1,8 +1,9 @@
 class ReflectionsController < ApplicationController
   include IfThenRulesHelper
   def create
-    rule = current_user.if_then_rules.find(params[:if_then_rule_id])
-
+    authorize Reflection
+    rule = policy_scope(IfThenRule).find(params[:if_then_rule_id])
+    authorize rule
     reflection = rule.reflections.find_or_create_by!(
       user: current_user,
       reflected_on: Date.current
@@ -12,13 +13,13 @@ class ReflectionsController < ApplicationController
     streams << turbo_stream.replace(
       "today_rules",
       partial: "if_then_rules/dash_boards/today_rules",
-      locals: { rules: current_user.if_then_rules }
+      locals: { rules: policy_scope(IfThenRule) }
     )
 
     streams << turbo_stream.replace(
       "today_progress",
       partial: "if_then_rules/dash_boards/today_progress",
-      locals: { progress: today_progress(current_user.if_then_rules) }
+      locals: { progress: today_progress(policy_scope(IfThenRule)) }
     )
 
     if current_user.all_rules_completed_today?
@@ -32,24 +33,26 @@ class ReflectionsController < ApplicationController
   end
 
   def index
-    @reflections = current_user.reflections.includes(:if_then_rule).order(reflected_on: :desc)
+    authorize Reflection
+    @reflections = policy_scope(Reflection).includes(:if_then_rule).order(reflected_on: :desc)
   end
 
   def destroy
-  reflection = current_user.reflections.find(params[:id])
+  reflection = policy_scope(Reflection).find(params[:id])
+  authorize reflection
   reflection.destroy!
     streams = []
 
     streams << turbo_stream.replace(
       "today_rules",
       partial: "if_then_rules/dash_boards/today_rules",
-      locals: { rules: current_user.if_then_rules }
+      locals: { rules: policy_scope(IfThenRule) }
     )
 
         streams << turbo_stream.replace(
       "today_progress",
       partial: "if_then_rules/dash_boards/today_progress",
-      locals: { progress: today_progress(current_user.if_then_rules) }
+      locals: { progress: today_progress(policy_scope(IfThenRule)) }
     )
   render turbo_stream: streams, notice: "チェックを取り消ししました"
   end
